cache: implement a partial file cache
This file cache implementation allows for one to read transparently
a network stream (implemented via HttpRandomAccessStream), via a stream-
like cache-backed implementation.

As ExoPlayer shows some really weird behavior with regards to seeks on
the input source, we use a class that alleviates the drawbacks of
seeking by keeping multiple input streams, and when seeking, either pick
one of those streams or create a new one, positioned at where we're
seeking. That is the function of RandomAccessStreamReplicator.

The RandomAccessStreamReplicator stream is then wrapped by a LazyRandom
AccessStream, which merely delays the creation of the stream as much as
possible. This somewhat simplifies the code, since we can deal with
streams directly, without worrying about the cost of creating it.

With the wrapped stream, we then create the cache-backed stream, and
provide access to it via a RandomAccessStreamView, which handles
concurrent access.