cache: finish partial cache implementation
Make streams open only when open() is called, not on creation. Now the
LazyRandomAccessStream delays opening as much as possible (i.e. when
any operation is called).

close() now may be called from other threads. Streams should handle this,
as not to block the closing operation. If they are closed during an
operation, this operation may raise a ClosedException.

ExhaustiveRandomAccessStreamReader is now FileCacheDownloader again,
and returns the cache backing file on success. Not sure if this is the
best way to do it, but working with streams is way harder than working
with files (CustomScaleSubsamplingView won't support anything but files/
URIs, GifDrawable breaks with streams).

UpdateManager and ImageSaveTask now work again.